[WIP] Clean code.
