+++ partner_relations_in_tab/tablib/__init__.py
@@ -1,5 +1,4 @@
 # -*- coding: utf-8 -*-
-# Copyright 2014-2018 Therp BV <https://therp.nl>.
+# Copyright 2018 Therp BV <https://therp.nl>.
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
-from . import tablib
-from . import model
+from .tab import Tab
